research_agent/pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.2.6",
    "langchain-openai>=1.0.2",
    "tavily-python>=0.7.13",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
]
[[tool.uv.index]]
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
